restructure web submodule
