+++ src/modules/osubot/web/index.js
@@ -1,0 +1,3 @@
+import api from './api'
+import res from './res'
+export { api, res }
